Add facility edit and delete functionality
Introduced EditFacilitypage form for editing facility details and integrated it into Facilitypage with new update and delete buttons. FacilityService now supports updating and deleting facilities. Facilitypage UI and logic updated to allow staff to edit or remove facilities, with member restrictions applied. PaymentForm and BookingService were refactored for improved payment handling. Minor UI and event handler improvements were made to Editpage and Facilitypage.

+++ Booking-app/Booking-app/Services/PaymentService.cs
@@ -16,7 +16,7 @@ public class PaymentService {
         _paymentCollection = db.GetCollection<Payment>("Payment");
     }
 
-    public void CreatePayment(ObjectId memberId, ObjectId bookingId, int amount, string paymentMethod,string paymentType )
+    public void CreatePayment(ObjectId memberId, ObjectId bookingId, decimal amount, string paymentMethod,string paymentType )
     {
         var payment = new Payment
         {
